# copilot_core/docs/openapi.yaml
openapi: 3.0.3
info:
  title: PilotSuite Styx Core API
  description: "\n# PilotSuite Styx Core API Documentation\n\nComprehensive REST API\
    \ documentation for the PilotSuite Styx Core home automation platform.\n\n## Architecture\n\
    \nPilotSuite Styx Core is a modular home automation backend built on Flask, providing:\n\
    \n- **Smart Home Integration**: Native Home Assistant integration with bi-directional\
    \ sync\n- **AI-Powered Automation**: Pattern mining, habitus learning, and predictive\
    \ automation\n- **Multi-Home Support**: Cross-home entity sharing and federated\
    \ learning\n- **Real-time Monitoring**: System health, energy tracking, and network\
    \ diagnostics\n- **Notification Engine**: Multi-channel notifications with digest\
    \ and scheduling\n\n## Authentication\n\nAll API endpoints require authentication\
    \ via one of the following methods:\n\n### API Key Authentication\nMost endpoints\
    \ use API Key authentication via the `X-API-Key` header:\n```\nX-API-Key: your-api-key-here\n\
    ```\n\n### Bearer Token Authentication\nSome endpoints (Notifications, Telegram)\
    \ use Bearer token authentication:\n```\nAuthorization: Bearer your-token-here\n\
    ```\n\n## Versioning\n\nAPI versioning is supported via the `Accept-Version` header:\n\
    ```\nAccept-Version: v1\n```\n\nDeprecated endpoints include `Deprecation` and\
    \ `Sunset` headers in responses.\n\n## Response Format\n\nAll responses follow\
...
    \ }\n}\n```\n"
...
servers:
...
security:
- apiKeyAuth: []
- bearerAuth: []
paths:
...
  /api/v1/tags:
    get:
      summary: List all tags
      description: Returns all tags.
      operationId: listTags
      tags:
      - Tags
      responses:
        '200':
          description: List of tags
      security:
      - bearerAuth: []
    post:
      summary: Create tag
      description: Creates a new tag.
      operationId: createTag
      tags:
      - Tags
      responses:
        '201':
          description: Tag created
      security:
      - bearerAuth: []
...
components:
...
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key authentication for most endpoints
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token authentication for Notifications and Telegram APIs